phone-number-validation/requirements.txt
Flask
mysql-connector-python
phonenumbers
